notifier/requirements.txt
requests==2.32.5
python-dotenv==1.0.0
requests==2.32.5
